Issue #150
CHANGE:
  - kvStoreInterfaces/redis.go Now houses common redis calls used
    **throughout** the application. The `server/peerStore/redisStore.go` file
    just implements a wrapper around these.
